+++ tests/post/steps/test_ingress.py
@@ -1,8 +1,11 @@
+import json
+import os
 import requests
 import requests.exceptions
 
 import pytest
 from pytest_bdd import given, parsers, scenario, then, when
+import testinfra
 
 from tests import utils
 
@@ -22,9 +25,28 @@ def test_access_http_services_on_control_plane_ip(host):
     pass
 
 
+@scenario("../features/ingress.feature", "Change Control Plane Ingress IP to node-1 IP")
+def test_change_cp_ingress_ip(host, teardown):
+    pass
+
+
 @pytest.fixture(scope="function")
 def context():
     return {}
+
+
+@pytest.fixture
+def teardown(context, host, ssh_config, version):
+    yield
+    if "bootstrap_to_restore" in context:
+        with host.sudo():
+            host.check_output(
+                "cp {} /etc/metalk8s/bootstrap.yaml".format(
+                    context["bootstrap_to_restore"]
+                )
+            )
+
+        re_configure_cp_ingress(host, version, ssh_config)
 
 
 @given("the node control-plane IP is not equal to its workload-plane IP")
@@ -67,6 +89,17 @@ def perform_request(host, context, protocol, port, plane):
         context["exception"] = exc
 
 
+@when(parsers.parse("we update control plane ingress IP to node '{node_name}' IP"))
+def update_cp_ingress_ip(host, context, ssh_config, version, node_name):
+    node = testinfra.get_host(node_name, ssh_config=ssh_config)
+    ip = utils.get_grain(node, "metalk8s:control_plane_ip")
+
+    bootstrap_patch = {"networks": {"controlPlane": {"ingress": {"ip": ip}}}}
+
+    patch_bootstrap_config(context, host, bootstrap_patch)
+    re_configure_cp_ingress(host, version, ssh_config)
+
+
 @then(
     parsers.re(r"the server returns (?P<status_code>\d+) '(?P<reason>.+)'"),
     converters=dict(status_code=int),
@@ -82,3 +115,48 @@ def server_returns(host, context, status_code, reason):
 def server_does_not_respond(host, context):
     assert "exception" in context
     assert isinstance(context["exception"], requests.exceptions.ConnectionError)
+
+
+@then(parsers.parse("the control plane ingress IP is equal to node '{node_name}' IP"))
+def check_cp_ingress_node_ip(control_plane_ingress_ip, node_name, ssh_config):
+    node = testinfra.get_host(node_name, ssh_config=ssh_config)
+    ip = utils.get_grain(node, "metalk8s:control_plane_ip")
+
+    assert control_plane_ingress_ip == ip
+
+
+def patch_bootstrap_config(context, host, patch):
+    with host.sudo():
+        cmd_ret = host.check_output("salt-call --out json --local temp.dir")
+
+    tmp_dir = json.loads(cmd_ret)["local"]
+
+    with host.sudo():
+        host.check_output("cp /etc/metalk8s/bootstrap.yaml {}".format(tmp_dir))
+
+    context["bootstrap_to_restore"] = os.path.join(tmp_dir, "bootstrap.yaml")
+
+    with host.sudo():
+        host.check_output(
+            "salt-call --local --retcode-passthrough state.single "
+            "file.serialize /etc/metalk8s/bootstrap.yaml "
+            "dataset='{}' "
+            "merge_if_exists=True".format(json.dumps(patch))
+        )
+
+
+def re_configure_cp_ingress(host, version, ssh_config):
+    with host.sudo():
+        host.check_output(
+            "salt-call --retcode-passthrough state.sls "
+            "metalk8s.kubernetes.apiserver saltenv=metalk8s-{}".format(version)
+        )
+
+    command = [
+        "salt-run",
+        "state.orchestrate",
+        "metalk8s.orchestrate.update-control-plane-ingress-ip",
+        "saltenv=metalk8s-{}".format(version),
+    ]
+
+    utils.run_salt_command(host, command, ssh_config)